feat(megathor): Added stand alone Ansible k8s deployment playbook
This adds a basic Ansible playbook for deploying Thorium and it's
dependencies on top of an existing locally hosted k8s cluster. As of this
commit, the playbook is not intented for use in hosted cloud environments
where managed databases, tracing, and/or ingress services might be used.
However, the roles within this playbook could be taken in part or whole
to deploy any missing Thorium dependencies.