fix(api): Fixed integration tests and incorrect client identify route
Fixed various issues with integration tests, including an incorrect test
config template path, scylla consistency issues, and lints. Also opted
to use std LazyLock instead of lazy_static and tokio OnceCell for
statics from futures, allowing for the removal of lazy_static and
async_once as dependencies. Tests also  now create S3 buckets with strict
1-day lifecycle policies if they don't already exist. Tests in GitLab
should now be in good shape.

Running tests again led to a discovery of a bad 'identify' route in the
client. 'identify' when mounted to '/' in axum is actually at '../api'
not '../api/' as unintuitive as that may appear.